+++ wasmerexec/wasmerexec.go
@@ -14,7 +14,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.WasmExit(args[0].I32())
+			mod.WasmExit(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -23,7 +23,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.WasmWrite(args[0].I32())
+			mod.WasmWrite(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -32,7 +32,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ResetMemoryDataView(args[0].I32())
+			mod.ResetMemoryDataView(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -41,7 +41,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.Nanotime1(args[0].I32())
+			mod.Nanotime1(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -50,7 +50,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.Walltime(args[0].I32())
+			mod.Walltime(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -59,7 +59,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ScheduleTimeoutEvent(args[0].I32())
+			mod.ScheduleTimeoutEvent(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -68,7 +68,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ClearTimeoutEvent(args[0].I32())
+			mod.ClearTimeoutEvent(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -77,7 +77,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.GetRandomData(args[0].I32())
+			mod.GetRandomData(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -86,7 +86,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.FinalizeRef(args[0].I32())
+			mod.FinalizeRef(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -95,7 +95,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.StringVal(args[0].I32())
+			mod.StringVal(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -104,7 +104,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueGet(args[0].I32())
+			mod.ValueGet(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -113,7 +113,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueSet(args[0].I32())
+			mod.ValueSet(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -122,7 +122,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueDelete(args[0].I32())
+			mod.ValueDelete(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -131,7 +131,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueIndex(args[0].I32())
+			mod.ValueIndex(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -140,7 +140,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueSetIndex(args[0].I32())
+			mod.ValueSetIndex(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -149,7 +149,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueCall(args[0].I32())
+			mod.ValueCall(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -158,7 +158,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueInvoke(args[0].I32())
+			mod.ValueInvoke(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -167,7 +167,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueNew(args[0].I32())
+			mod.ValueNew(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -176,7 +176,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueLength(args[0].I32())
+			mod.ValueLength(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -185,7 +185,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValuePrepareString(args[0].I32())
+			mod.ValuePrepareString(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -194,7 +194,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueLoadString(args[0].I32())
+			mod.ValueLoadString(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -203,7 +203,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.ValueInstanceOf(args[0].I32())
+			mod.ValueInstanceOf(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -212,7 +212,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.CopyBytesToGo(args[0].I32())
+			mod.CopyBytesToGo(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -221,7 +221,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.CopyBytesToJS(args[0].I32())
+			mod.CopyBytesToJS(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
@@ -230,7 +230,7 @@ func Import(store *wasmer.Store, instance wasmexec.Instance) (*wasmer.ImportObje
 		store,
 		wasmer.NewFunctionType(wasmer.NewValueTypes(wasmer.I32), wasmer.NewValueTypes()),
 		func(args []wasmer.Value) ([]wasmer.Value, error) {
-			mod.Debug(args[0].I32())
+			mod.Debug(uint32(args[0].I32()))
 			return []wasmer.Value{}, nil
 		},
 	)
